requirements.txt
PyMuPDF==1.24.2
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.46
langchain_openai==0.1.4
langchain_text_splitters==0.0.1
loguru==0.7.2
openai==1.24.0
python-dotenv==1.0.1
Requests==2.31.0
mongoengine==0.28.2
docx2txt==0.8
uvicorn==0.30.5
fastapi==0.112.0
python-multipart==0.0.9
pydantic==2.8.2
llama-index-core==0.10.60
llama-index-readers-file==0.1.32
llama-parse==0.4.9
async-timeout==4.0.3